requirements.txt
click==8.1.3
Flask==2.2.5
imageio==2.28.1
importlib-metadata==6.6.0

numpy==1.21.6
packaging==23.1
Pillow==9.5.0
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.7.3

gunicorn==20.1.0
tensorflow